# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Founder Brain Audit",
    "description": "Decision intelligence from your AI conversations.",
    "version": "0.1.0"
  },
  "paths": {
    "/upload": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Upload File",
        "operationId": "upload_file_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/entities/sources": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "List Sources",
        "operationId": "list_sources_entities_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationSourceOut"
                  },
                  "type": "array",
                  "title": "Response List Sources Entities Sources Get"
                }
              }
            }
          }
        }
      }
    },
    "/entities/sources/{source_id}": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Get Source",
        "description": "Poll extraction status for a specific source after upload.",
        "operationId": "get_source_entities_sources__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationSourceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/entities/decisions": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "List Decisions",
        "operationId": "list_decisions_entities_decisions_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.0,
              "title": "Min Confidence"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionOut"
                  },
                  "title": "Response List Decisions Entities Decisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/entities/goals": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "List Goals",
        "operationId": "list_goals_entities_goals_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoalOut"
                  },
                  "title": "Response List Goals Entities Goals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/entities/constraints": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "List Constraints",
        "operationId": "list_constraints_entities_constraints_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConstraintOut"
                  },
                  "title": "Response List Constraints Entities Constraints Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/entities/open-questions": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "List Open Questions",
        "operationId": "list_open_questions_entities_open_questions_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpenQuestionOut"
                  },
                  "title": "Response List Open Questions Entities Open Questions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/entities/action-items": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "List Action Items",
        "operationId": "list_action_items_entities_action_items_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActionItemOut"
                  },
                  "title": "Response List Action Items Entities Action Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/insights": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Get Insights",
        "operationId": "get_insights_insights_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightReport"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionItemOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/ActionStatus"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "supporting_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supporting Snippet"
          }
        },
        "type": "object",
        "required": [
          "id",
          "description",
          "status",
          "confidence",
          "source_reference",
          "supporting_snippet"
        ],
        "title": "ActionItemOut"
      },
      "ActionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "done",
          "cancelled",
          "unknown"
        ],
        "title": "ActionStatus"
      },
      "BlindSpot": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "discussion_count": {
            "type": "integer",
            "title": "Discussion Count"
          },
          "action_count": {
            "type": "integer",
            "title": "Action Count"
          },
          "ratio": {
            "type": "number",
            "title": "Ratio"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "discussion_count",
          "action_count",
          "ratio"
        ],
        "title": "BlindSpot"
      },
      "Body_upload_file_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_upload_post"
      },
      "ConstraintOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "supporting_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supporting Snippet"
          }
        },
        "type": "object",
        "required": [
          "id",
          "description",
          "confidence",
          "source_reference",
          "supporting_snippet"
        ],
        "title": "ConstraintOut"
      },
      "ConversationSourceOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          },
          "conversation_count": {
            "type": "integer",
            "title": "Conversation Count"
          },
          "extraction_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Status"
          },
          "provider_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Used"
          },
          "entities_extracted": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entities Extracted"
          },
          "extraction_confidence_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Confidence Avg"
          },
          "extraction_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "source_type",
          "uploaded_at",
          "conversation_count"
        ],
        "title": "ConversationSourceOut"
      },
      "DecisionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "supporting_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supporting Snippet"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonOut"
            },
            "type": "array",
            "title": "Reasons",
            "default": []
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceOut"
            },
            "type": "array",
            "title": "Evidence",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "timestamp",
          "confidence",
          "source_reference",
          "supporting_snippet",
          "source_id"
        ],
        "title": "DecisionOut"
      },
      "DecisionReversal": {
        "properties": {
          "original": {
            "$ref": "#/components/schemas/DecisionOut"
          },
          "reversal": {
            "$ref": "#/components/schemas/DecisionOut"
          },
          "similarity": {
            "type": "number",
            "title": "Similarity"
          }
        },
        "type": "object",
        "required": [
          "original",
          "reversal",
          "similarity"
        ],
        "title": "DecisionReversal"
      },
      "EvidenceOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          }
        },
        "type": "object",
        "required": [
          "id",
          "description",
          "source_reference"
        ],
        "title": "EvidenceOut"
      },
      "FallbackStep": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "status"
        ],
        "title": "FallbackStep"
      },
      "GoalOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/GoalStatus"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "supporting_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supporting Snippet"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "frequency": {
            "type": "integer",
            "title": "Frequency"
          }
        },
        "type": "object",
        "required": [
          "id",
          "description",
          "status",
          "confidence",
          "source_reference",
          "supporting_snippet",
          "timestamp",
          "frequency"
        ],
        "title": "GoalOut"
      },
      "GoalStatus": {
        "type": "string",
        "enum": [
          "open",
          "achieved",
          "abandoned",
          "unknown"
        ],
        "title": "GoalStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InsightReport": {
        "properties": {
          "recurring_questions": {
            "items": {
              "$ref": "#/components/schemas/RecurringQuestionGroup"
            },
            "type": "array",
            "title": "Recurring Questions"
          },
          "decision_reversals": {
            "items": {
              "$ref": "#/components/schemas/DecisionReversal"
            },
            "type": "array",
            "title": "Decision Reversals"
          },
          "top_goals": {
            "items": {
              "$ref": "#/components/schemas/GoalOut"
            },
            "type": "array",
            "title": "Top Goals"
          },
          "blind_spots": {
            "items": {
              "$ref": "#/components/schemas/BlindSpot"
            },
            "type": "array",
            "title": "Blind Spots"
          },
          "total_decisions": {
            "type": "integer",
            "title": "Total Decisions"
          },
          "total_open_questions": {
            "type": "integer",
            "title": "Total Open Questions"
          },
          "total_action_items": {
            "type": "integer",
            "title": "Total Action Items"
          }
        },
        "type": "object",
        "required": [
          "recurring_questions",
          "decision_reversals",
          "top_goals",
          "blind_spots",
          "total_decisions",
          "total_open_questions",
          "total_action_items"
        ],
        "title": "InsightReport"
      },
      "OpenQuestionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "supporting_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supporting Snippet"
          }
        },
        "type": "object",
        "required": [
          "id",
          "description",
          "confidence",
          "source_reference",
          "supporting_snippet"
        ],
        "title": "OpenQuestionOut"
      },
      "ReasonOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "description",
          "confidence"
        ],
        "title": "ReasonOut"
      },
      "RecurringQuestionGroup": {
        "properties": {
          "representative": {
            "type": "string",
            "title": "Representative"
          },
          "occurrences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Occurrences"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "representative",
          "occurrences",
          "count"
        ],
        "title": "RecurringQuestionGroup"
      },
      "UploadResponse": {
        "properties": {
          "source_id": {
            "type": "integer",
            "title": "Source Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "conversation_count": {
            "type": "integer",
            "title": "Conversation Count"
          },
          "entities_extracted": {
            "type": "integer",
            "title": "Entities Extracted"
          },
          "provider_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Used"
          },
          "extraction_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Status"
          },
          "extraction_confidence_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Confidence Avg"
          },
          "extraction_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Duration Ms"
          },
          "fallback_chain": {
            "items": {
              "$ref": "#/components/schemas/FallbackStep"
            },
            "type": "array",
            "title": "Fallback Chain",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "filename",
          "source_type",
          "conversation_count",
          "entities_extracted"
        ],
        "title": "UploadResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}